cheri: Fix compile/warnings in drivers/media/
- The UAPI has to change in a view places for CHERI.
- Skip CONFIG_DVB_VIDTV and CONFIG_VIDEO_MEDIATEK_*. This have
  embedded pointers in structs apparently used for mailbox
  communication.
- Otherwise only standard uintptr_t vs. unsigned long issues.

+++ drivers/media/platform/chips-media/wave5/wave5-vdi.c
@@ -198,7 +198,7 @@ void wave5_vdi_free_sram(struct vpu_device *vpu_dev)
 		return;
 
 	if (vb->vaddr)
-		gen_pool_free(vpu_dev->sram_pool, (unsigned long)vb->vaddr,
+		gen_pool_free(vpu_dev->sram_pool, (uintptr_t)vb->vaddr,
 			      vb->size);
 
 	memset(vb, 0, sizeof(*vb));
